feat: add input for regexp
