+++ src/components/Chat/PublishModal/TargetAudienceFilterTypeSelect.tsx
@@ -1,0 +1,70 @@
+import { IconChevronDown } from '@tabler/icons-react';
+import { MouseEvent, useState } from 'react';
+
+import { useTranslation } from 'next-i18next';
+
+import classNames from 'classnames';
+
+import { FiltersTypes } from '@/src/types/share';
+import { Translation } from '@/src/types/translation';
+
+import { Menu, MenuItem } from '../../Common/DropdownMenu';
+
+interface FilterTypeProps {
+  id: string;
+  filterTypes: string[];
+  selectedType: string;
+  onChangeFilterType: (filterType: FiltersTypes) => void;
+}
+
+export function FilterTypeSelect({
+  id,
+  filterTypes,
+  selectedType,
+  onChangeFilterType,
+}: FilterTypeProps) {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const { t } = useTranslation(Translation.SideBar);
+
+  const onChangeHandler = (e: MouseEvent<HTMLButtonElement>) => {
+    onChangeFilterType(e.currentTarget.value as FiltersTypes);
+    setIsOpen(false);
+  };
+
+  return (
+    <div
+      data-qa={`filter-selector-${id}`}
+      className="h-[38px] w-full max-w-[125px] grow rounded border border-primary text-xs focus-within:border-accent-primary focus:border-accent-primary"
+    >
+      <Menu
+        className="w-full px-3"
+        onOpenChange={setIsOpen}
+        trigger={
+          <div className="flex items-center justify-between gap-2">
+            {selectedType}
+            <IconChevronDown
+              data-qa={`open-filter-dropdown-${id}`}
+              className={classNames(
+                'shrink-0 text-primary transition-all',
+                isOpen && 'rotate-180',
+              )}
+              width={18}
+              height={18}
+            />
+          </div>
+        }
+      >
+        {filterTypes.map((filterType) => (
+          <MenuItem
+            key={filterType}
+            className="max-w-[350px] text-xs hover:bg-accent-primary-alpha"
+            item={t(filterType)}
+            value={filterType}
+            onClick={onChangeHandler}
+          />
+        ))}
+      </Menu>
+    </div>
+  );
+}
